requirements.txt
Flask
pymongo
redis
textblob
python-dotenv
